Django Project/heart_prediction/requirements.txt
absl-py==0.11.0
aiohttp==3.7.3
alabaster==0.7.12
alembic==1.4.1
argh==0.26.2
argon2-cffi==20.1.0
asgiref==3.2.10
astroid==2.4.2
astropy==4.2
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==20.3.0
autopep8==1.5.4
Babel==2.9.0
backcall==0.2.0
bcrypt==3.2.0
bleach==3.2.1
blinker==1.4
bokeh==2.2.3
Bottleneck==1.3.2
brotlipy==0.7.0
cachetools==4.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
click==7.1.2
cloudpickle==1.6.0
colorama==0.4.4
comet-ml==3.3.2
configobj==5.0.6
coverage==5.4
cryptography==3.3.1
cycler==0.10.0
cytoolz==0.11.0
dask==2021.1.1
databricks-cli==0.14.3
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20200713
dill==0.3.3
distributed==2021.1.1
dj-database-url==0.5.0
Django==3.0.9
django-cors-headers==3.5.0
django-filter==2.4.0
django-templated-mail==1.1.1
djangorestframework==3.11.0
djoser==2.0.5
docker==5.0.0
docutils==0.16
dotmap==1.3.23
dulwich==0.20.18
entrypoints==0.3
everett==1.0.3
fast-histogram==0.9
flake8==3.8.4
Flask==2.0.1
fsspec==0.8.3
future==0.18.2
gast==0.3.3
gitdb==4.0.7
GitPython==3.1.17
glue-core==0.15.6
glue-vispy-viewers==0.12.2
glueviz==1.0.0
google-auth==1.24.0
google-auth-oauthlib
google-pasta==0.2.0
graphviz==0.16
greenlet==1.1.0
grpcio==1.31.0
gunicorn==20.1.0
h5py==2.10.0
HeapDict==1.0.1
heroku==0.1.4
httpie==2.4.0
idna==2.10.0
imagecodecs==2020.5.30
imageio==2.9.0
imagesize==1.2.0
importlib-metadata
intervaltree==3.1.0
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils
ipywidgets==7.6.3
isort==5.6.4
itsdangerous==2.0.1
jedi==0.17.1
Jinja2==3.0.1
joblib==1.0.0
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.7.0
jupyterlab-pygments
jupyterlab-widgets==1.0.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==21.4.0
kiwisolver==1.3.0
lazy-object-proxy==1.4.3
locket==0.2.1
Mako==1.1.4
Markdown==3.3.3
MarkupSafe==2.0.1
matplotlib==3.3.2
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.2.0
mkl-random==1.2.0
mkl-service==2.3.0
mlflow==1.17.0
mysqlclient==2.0.3
netifaces==0.10.9
notebook==6.4.0
numpy==1.19.5
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
olefile==0.46
opencv-python-headless==4.5.2.54
opt-einsum==3.1.0
parso==0.7.0
partd==1.1.0
pathtools==0.1.2
pluggy==0.13.1
prometheus-client==0.10.1
prometheus-flask-exporter==0.18.2
protobuf==3.13.0
psycopg2==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pydot==1.4.1
pydotplus==2.0.2
pyflakes==2.2.0
PyJWT==2.1.0
pylint-django==2.2.0
pylint-plugin-utils==0.6
PyOpenGL==3.1.1a1
pyparsing==2.4.7
pyreadline==2.1
python-editor==1.0.4
python3-openid==3.2.0
pywin32==227
pywinpty==1.1.1
PyYAML==5.3.1
pyzmq==20.0.0
QDarkStyle==2.8.1
QtPy==1.9.0
querystring-parser==1.2.4
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
retrying==1.3.3

Rtree==0.9.4
scikit-image==0.17.2

seaborn==0.10.1
Send2Trash==1.5.0
smmap==4.0.0
snowballstemmer==2.0.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0

sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
SQLAlchemy==1.4.15
sqlparse==0.3.1
tabulate==0.8.9
tensorboard-plugin-wit==1.6.0
tensorflow==2.3.0
termcolor==1.1.0
terminado==0.10.0
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2021.1.14
toml==0.10.1
toolz==0.11.1
tornado==6.1
traitlets==5.0.5
typed-ast==1.4.1
typing-extensions==3.7.4.3
ujson==1.35
urllib3==1.25.11
waitress==2.0.0
watchdog==0.10.4
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==2.0.1
whitenoise==5.2.0
widgetsnbextension==3.5.1
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.11.2
wurlitzer==2.0.1
xgboost==1.3.3
xlrd==1.2.0
xlwt==1.3.0
yapf==0.30.0
yarl==1.5.1
zict==2.0.0
zipp ==3.4.0
